Use dynamic plugin for oauth config details page
