+++ frontend/packages/console-app/package.json
@@ -48,7 +48,8 @@
       "storageProvisioners": "src/components/storage/StorageClassProviders",
       "storageProvisionerDocs": "src/components/storage/Documentation",
       "nodeStatus": "src/components/nodes/status",
-      "nodeActions": "src/components/nodes/menu-actions.tsx"
+      "nodeActions": "src/components/nodes/menu-actions.tsx",
+      "oauthConfigDetailsPage": "src/components/oauth-config/OAuthConfigDetailsPage.tsx"
     }
   }
 }
