+++ frontend/packages/console-app/src/components/oauth-config/OAuthConfigDetails.tsx
@@ -1,7 +1,4 @@
 import * as React from 'react';
-import * as _ from 'lodash-es';
-import { useTranslation } from 'react-i18next';
-
 import { formatPrometheusDuration } from '@openshift-console/plugin-shared/src/datetime/prometheus';
 import { Alert } from '@patternfly/react-core';
 import {
@@ -10,76 +7,25 @@ import {
   DropdownToggle as DropdownToggleDeprecated,
 } from '@patternfly/react-core/deprecated';
 import { CaretDownIcon } from '@patternfly/react-icons/dist/esm/icons/caret-down-icon';
+import * as _ from 'lodash';
+import { useTranslation } from 'react-i18next';
 import { Link, useNavigate } from 'react-router-dom-v5-compat';
-import { useQueryParams } from '@console/shared';
-import { ClusterOperatorModel, OAuthModel } from '../../models';
-import { IdentityProvider, OAuthKind, referenceForModel } from '../../module/k8s';
-import { DetailsPage } from '../factory';
 import {
-  EmptyBox,
-  Kebab,
-  navFactory,
   resourcePathFromModel,
   ResourceSummary,
   SectionHeading,
-} from '../utils';
-
-const { common } = Kebab.factory;
-const menuActions = [...Kebab.getExtensionsActionsForKind(OAuthModel), ...common];
-
-const oAuthReference = referenceForModel(OAuthModel);
+} from '@console/internal/components/utils';
+import { ClusterOperatorModel } from '@console/internal/models';
+import { OAuthKind } from '@console/internal/module/k8s';
+import { IDP_TYPES } from '@console/shared/src/constants/auth';
+import { useQueryParams } from '@console/shared/src/hooks/useQueryParams';
+import { IdentityProviders } from './IdentityProviders';
 
 // Convert to ms for formatPrometheusDuration
 const tokenDuration = (seconds: number) =>
   _.isNil(seconds) ? '-' : formatPrometheusDuration(seconds * 1000);
 
-const IdentityProviders: React.FC<IdentityProvidersProps> = ({ identityProviders }) => {
-  const { t } = useTranslation();
-  return _.isEmpty(identityProviders) ? (
-    <EmptyBox label={t('public~Identity providers')} />
-  ) : (
-    <div className="co-table-container">
-      <table className="pf-v5-c-table pf-m-compact pf-m-border-rows">
-        <thead className="pf-v5-c-table__thead">
-          <tr className="pf-v5-c-table__tr">
-            <th className="pf-v5-c-table__th">{t('public~Name')}</th>
-            <th className="pf-v5-c-table__th">{t('public~Type')}</th>
-            <th className="pf-v5-c-table__th">{t('public~Mapping method')}</th>
-          </tr>
-        </thead>
-        <tbody className="pf-v5-c-table__tbody">
-          {_.map(identityProviders, (idp) => (
-            <tr className="pf-v5-c-table__tr" key={idp.name}>
-              <td className="pf-v5-c-table__td" data-test-idp-name={idp.name}>
-                {idp.name}
-              </td>
-              <td className="pf-v5-c-table__td" data-test-idp-type-for={idp.name}>
-                {idp.type}
-              </td>
-              <td className="pf-v5-c-table__td" data-test-idp-mapping-for={idp.name}>
-                {idp.mappingMethod || 'claim'}
-              </td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </div>
-  );
-};
-
-export const addIDPItems = Object.freeze({
-  basicauth: 'Basic Authentication',
-  github: 'GitHub',
-  gitlab: 'GitLab',
-  google: 'Google',
-  htpasswd: 'HTPasswd',
-  keystone: 'Keystone',
-  ldap: 'LDAP',
-  oidconnect: 'OpenID Connect',
-  requestheader: 'Request Header',
-});
-
-const OAuthDetails: React.FC<OAuthDetailsProps> = ({ obj }: { obj: OAuthKind }) => {
+export const OAuthConfigDetails: React.FC<OAuthDetailsProps> = ({ obj }: { obj: OAuthKind }) => {
   const navigate = useNavigate();
   const [isIDPOpen, setIDPOpen] = React.useState(false);
   const { identityProviders, tokenConfig } = obj.spec;
@@ -112,7 +58,7 @@ const OAuthDetails: React.FC<OAuthDetailsProps> = ({ obj }: { obj: OAuthKind }) 
     }
   };
 
-  const IDPDropdownItems = Object.entries(addIDPItems).map((idp) => {
+  const IDPDropdownItems = Object.entries(IDP_TYPES).map((idp) => {
     const [key, value] = idp;
 
     return (
@@ -188,19 +134,6 @@ const OAuthDetails: React.FC<OAuthDetailsProps> = ({ obj }: { obj: OAuthKind }) 
       </div>
     </>
   );
-};
-
-export const OAuthDetailsPage: React.FC = (props) => (
-  <DetailsPage
-    {...props}
-    kind={oAuthReference}
-    menuActions={menuActions}
-    pages={[navFactory.details(OAuthDetails), navFactory.editYaml()]}
-  />
-);
-
-type IdentityProvidersProps = {
-  identityProviders: IdentityProvider[];
 };
 
 type OAuthDetailsProps = {
