+++ frontend/packages/console-shared/src/constants/auth.ts
@@ -1,0 +1,11 @@
+export const IDP_TYPES = Object.freeze({
+  basicauth: 'Basic Authentication',
+  github: 'GitHub',
+  gitlab: 'GitLab',
+  google: 'Google',
+  htpasswd: 'HTPasswd',
+  keystone: 'Keystone',
+  ldap: 'LDAP',
+  oidconnect: 'OpenID Connect',
+  requestheader: 'Request Header',
+});
