+++ frontend/public/locales/en/public.json
@@ -241,23 +241,6 @@
   "Email": "Email",
   "The list of attributes whose values should be used as the email address.": "The list of attributes whose values should be used as the email address.",
   "More options": "More options",
-  "Identity providers": "Identity providers",
-  "Mapping method": "Mapping method",
-  "Basic Authentication": "Basic Authentication",
-  "GitHub": "GitHub",
-  "GitLab": "GitLab",
-  "Google": "Google",
-  "HTPasswd": "HTPasswd",
-  "Keystone": "Keystone",
-  "LDAP": "LDAP",
-  "OpenID Connect": "OpenID Connect",
-  "Request Header": "Request Header",
-  "OAuth details": "OAuth details",
-  "Access token max age": "Access token max age",
-  "Identity providers determine how users log into the cluster.": "Identity providers determine how users log into the cluster.",
-  "New identity provider added.": "New identity provider added.",
-  "Authentication is being reconfigured. The new identity provider will be available once reconfiguration is complete.": "Authentication is being reconfigured. The new identity provider will be available once reconfiguration is complete.",
-  "View authentication conditions for reconfiguration status.": "View authentication conditions for reconfiguration status.",
   "Add Identity Provider: OpenID Connect": "Add Identity Provider: OpenID Connect",
   "Integrate with an OpenID Connect identity provider using an Authorization Code Flow.": "Integrate with an OpenID Connect identity provider using an Authorization Code Flow.",
   "Issuer URL": "Issuer URL",
@@ -1850,6 +1833,8 @@
   "ClusterRoles": "ClusterRoles",
   "SelfSubjectAccessReview": "SelfSubjectAccessReview",
   "SelfSubjectAccessReviews": "SelfSubjectAccessReviews",
+  "SelfSubjectReview": "SelfSubjectReview",
+  "SelfSubjectReviews": "SelfSubjectReviews",
   "ResourceAccessReview": "ResourceAccessReview",
   "ResourceAccessReviews": "ResourceAccessReviews",
   "LocalResourceAccessReview": "LocalResourceAccessReview",
@@ -1969,6 +1954,23 @@
   "Select a log file above": "Select a log file above",
   "Filter by unit": "Filter by unit",
   "Unit": "Unit",
+  "Identity providers": "Identity providers",
+  "Mapping method": "Mapping method",
+  "Basic Authentication": "Basic Authentication",
+  "GitHub": "GitHub",
+  "GitLab": "GitLab",
+  "Google": "Google",
+  "HTPasswd": "HTPasswd",
+  "Keystone": "Keystone",
+  "LDAP": "LDAP",
+  "OpenID Connect": "OpenID Connect",
+  "Request Header": "Request Header",
+  "OAuth details": "OAuth details",
+  "Access token max age": "Access token max age",
+  "Identity providers determine how users log into the cluster.": "Identity providers determine how users log into the cluster.",
+  "New identity provider added.": "New identity provider added.",
+  "Authentication is being reconfigured. The new identity provider will be available once reconfiguration is complete.": "Authentication is being reconfigured. The new identity provider will be available once reconfiguration is complete.",
+  "View authentication conditions for reconfiguration status.": "View authentication conditions for reconfiguration status.",
   "Loading {{title}} status": "Loading {{title}} status",
   "graph timespan": "graph timespan",
   "Reset zoom": "Reset zoom",
